Refactor UserInfo block a bit
