+++ src/components/UserInfo/UserInfo.jsx
@@ -1,77 +1,12 @@
 import React, { PropTypes } from 'react';
-import { connect } from 'react-redux';
-import SignIn from '../SignIn/SignIn';
-import SignOut from '../SignOut/SignOut';
-import SignUp from '../SignUp/SignUp';
 import './UserInfo.sass';
 
-class UserInfo extends React.Component {
-  constructor() {
-    super();
-
-    this.toggleTabHandler = this.toggleTabHandler.bind(this);
-    this.state = { tab: 'SIGN_IN' };
-  }
-
-  toggleTabHandler(tab) {
-    return (function toggleTab() {
-      this.setState({ tab });
-    }).bind(this);
-  }
-
-  userInfo() {
-    // TODO: Add actual info
-    const user = this.props.user; // eslint-disable-line no-unused-vars
-    return (
-      <div className="UserInfo__Main">
-        <SignOut />
-      </div>
-    );
-  }
-
-  signUpInBlock() {
-    const tab = this.state.tab;
-    let signUpIn;
-    switch (tab) {
-      case 'SIGN_IN':
-        signUpIn = <SignIn />;
-        break;
-      case 'SIGN_UP':
-        signUpIn = <SignUp />;
-        break;
-      default:
-    }
-
-    return (
-      <div className="UserInfo">
-        <ul className="UserInfo__Tabs">
-          <li className={`UserInfo__Tab${tab === 'SIGN_IN' ? '--Active' : ''}`}>
-            <a tabIndex="0" href="javascript:;" onClick={this.toggleTabHandler('SIGN_IN')}>Sign In</a>
-          </li>
-          <li className={`UserInfo__Tab${tab === 'SIGN_UP' ? '--Active' : ''}`}>
-            <a tabIndex="0" href="javascript:;" onClick={this.toggleTabHandler('SIGN_UP')}>Sign Up</a>
-          </li>
-        </ul>
-        {signUpIn}
-      </div>
-    );
-  }
-
-  render() {
-    return (
-      <div className="UserInfo">
-        {this.props.user ? this.userInfo() : this.signUpInBlock()}
-      </div>
-    );
-  }
+export default function UserInfo(props) {
+  return (
+    <div className="UserInfo">
+      <p className="UserInfo__Line">Logged in as: {props.user.login}</p>
+    </div>
+  );
 }
 
 UserInfo.propTypes = { user: PropTypes.objectOf(PropTypes.string) };
-
-function mapStateToProps({ auth }) {
-  let user = auth.get('user');
-  if (user) user = user.toJS();
-  return { user };
-}
-
-export default connect(mapStateToProps)(UserInfo);
